Fix issue where deleting interface caused crash
The main fix was that we were trying to free the vif_ctx twice. This caused the kernel to run into an error later on. This was fixed by removing the "nrf_fmac_vif_clear_ctx" line from the delete virtual interface callback.

Also added a mem debug which needs to be adjusted. There's also a lot of debug lines that need to be removed, but for now its working correctly.